requirements.txt
aiohttp==3.9.5
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
gunicorn
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
docutils==0.21.2
et-xmlfile==1.1.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lockfile==0.12.2
lxml==5.2.2
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.1.1
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pid==3.0.4
psutil==6.0.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
selenium==4.22.0
setproctitle==1.3.3
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.4
# Removed logging==0.4.9.6
# Removed duplicate pandas and flask
# Removed bs4==0.0.2
# Removed flask (duplicate)
# Removed pandas (duplicate)
